PIM-3564 : move saver and remover interfaces
