+++ spec/Pim/Bundle/CatalogBundle/Manager/AttributeGroupManagerSpec.php
@@ -22,17 +22,17 @@ class AttributeGroupManagerSpec extends ObjectBehavior
 
     function it_is_a_saver()
     {
-        $this->shouldImplement('Pim\Component\Resource\Model\SaverInterface');
+        $this->shouldImplement('Akeneo\Component\Persistence\SaverInterface');
     }
 
     function it_is_a_bulk_saver()
     {
-        $this->shouldImplement('Pim\Component\Resource\Model\BulkSaverInterface');
+        $this->shouldImplement('Akeneo\Component\Persistence\BulkSaverInterface');
     }
 
     function it_is_a_remover()
     {
-        $this->shouldImplement('Pim\Component\Resource\Model\RemoverInterface');
+        $this->shouldImplement('Akeneo\Component\Persistence\RemoverInterface');
     }
 
     function it_throws_an_exception_when_removing_an_attribute_group_and_the_default_group_does_not_exist(
