+++ spec/Pim/Bundle/DataGridBundle/Manager/DatagridViewManagerSpec.php
@@ -20,12 +20,12 @@ class DatagridViewManagerSpec extends ObjectBehavior
 
     function it_is_a_saver()
     {
-        $this->shouldImplement('Pim\Component\Resource\Model\SaverInterface');
+        $this->shouldImplement('Akeneo\Component\Persistence\SaverInterface');
     }
 
     function it_is_a_remover()
     {
-        $this->shouldImplement('Pim\Component\Resource\Model\RemoverInterface');
+        $this->shouldImplement('Akeneo\Component\Persistence\RemoverInterface');
     }
 
     function it_throws_exception_when_save_anything_else_than_a_view()
