+++ spec/Pim/Bundle/EnrichBundle/Form/Handler/BaseHandlerSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\EnrichBundle\Form\Handler;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Pim\Component\Resource\Model\SaverInterface;
+use Akeneo\Component\Persistence\SaverInterface;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
