+++ src/Akeneo/Component/Persistence/BulkRemoverInterface.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Pim\Component\Resource\Model;
+namespace Akeneo\Component\Persistence;
 
 /**
- * Remover interface, provides a minimal contract to remove a single business object
+ * Bulk remover interface, provides a minimal contract to delete many business objects
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface RemoverInterface
+interface BulkRemoverInterface
 {
     /**
-     * Delete a single object
+     * Delete many objects
      *
-     * @param mixed $object  The object to delete
+     * @param array $objects The objects to delete
      * @param array $options The delete options
      *
      * @throws \InvalidArgumentException
      */
-    public function remove($object, array $options = []);
+    public function removeAll(array $objects, $options = []);
 }
