+++ src/Pim/Bundle/CatalogBundle/Saver/BaseSaver.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\CatalogBundle\Saver;
 
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Common\Util\ClassUtils;
-use Pim\Component\Resource\Model\SaverInterface;
+use Akeneo\Component\Persistence\SaverInterface;
 
 /**
  * Base saver, declared as different services for different classes
